+++ site/src/app/routeTree.gen.ts
@@ -14,6 +14,7 @@ import { Route as PageRouteImport } from "./routes/_page";
 import { Route as PageDemosRouteImport } from "./routes/_page/demos";
 import { Route as PageConnectorsIndexRouteImport } from "./routes/_page/connectors/index";
 import { Route as PageConnectorsAjvRouteImport } from "./routes/_page/connectors/ajv";
+import { Route as PageConnectorsContenteditableRouteImport } from "./routes/_page/connectors/contenteditable";
 import { Route as PageConnectorsReactRouteImport } from "./routes/_page/connectors/react";
 import { Route as PageConnectorsReactHookFormRouteImport } from "./routes/_page/connectors/react-hook-form";
 import { Route as PageConnectorsTanstackTableRouteImport } from "./routes/_page/connectors/tanstack-table";
@@ -63,6 +64,12 @@ const PageConnectorsAjvRoute = PageConnectorsAjvRouteImport.update({
   path: "/connectors/ajv",
   getParentRoute: () => PageRoute,
 } as any);
+const PageConnectorsContenteditableRoute =
+  PageConnectorsContenteditableRouteImport.update({
+    id: "/connectors/contenteditable",
+    path: "/connectors/contenteditable",
+    getParentRoute: () => PageRoute,
+  } as any);
 const PageConnectorsReactRoute = PageConnectorsReactRouteImport.update({
   id: "/connectors/react",
   path: "/connectors/react",
@@ -191,6 +198,7 @@ export interface FileRoutesByFullPath {
   "/": typeof IndexRoute;
   "/demos": typeof PageDemosRoute;
   "/connectors/ajv": typeof PageConnectorsAjvRoute;
+  "/connectors/contenteditable": typeof PageConnectorsContenteditableRoute;
   "/connectors/react": typeof PageConnectorsReactRoute;
   "/connectors/react-hook-form": typeof PageConnectorsReactHookFormRoute;
   "/connectors/tanstack-table": typeof PageConnectorsTanstackTableRoute;
@@ -221,6 +229,7 @@ export interface FileRoutesByTo {
   "/": typeof IndexRoute;
   "/demos": typeof PageDemosRoute;
   "/connectors/ajv": typeof PageConnectorsAjvRoute;
+  "/connectors/contenteditable": typeof PageConnectorsContenteditableRoute;
   "/connectors/react": typeof PageConnectorsReactRoute;
   "/connectors/react-hook-form": typeof PageConnectorsReactHookFormRoute;
   "/connectors/tanstack-table": typeof PageConnectorsTanstackTableRoute;
@@ -253,6 +262,7 @@ export interface FileRoutesById {
   "/_page": typeof PageRouteWithChildren;
   "/_page/demos": typeof PageDemosRoute;
   "/_page/connectors/ajv": typeof PageConnectorsAjvRoute;
+  "/_page/connectors/contenteditable": typeof PageConnectorsContenteditableRoute;
   "/_page/connectors/react": typeof PageConnectorsReactRoute;
   "/_page/connectors/react-hook-form": typeof PageConnectorsReactHookFormRoute;
   "/_page/connectors/tanstack-table": typeof PageConnectorsTanstackTableRoute;
@@ -285,6 +295,7 @@ export interface FileRouteTypes {
     | "/"
     | "/demos"
     | "/connectors/ajv"
+    | "/connectors/contenteditable"
     | "/connectors/react"
     | "/connectors/react-hook-form"
     | "/connectors/tanstack-table"
@@ -315,6 +326,7 @@ export interface FileRouteTypes {
     | "/"
     | "/demos"
     | "/connectors/ajv"
+    | "/connectors/contenteditable"
     | "/connectors/react"
     | "/connectors/react-hook-form"
     | "/connectors/tanstack-table"
@@ -346,6 +358,7 @@ export interface FileRouteTypes {
     | "/_page"
     | "/_page/demos"
     | "/_page/connectors/ajv"
+    | "/_page/connectors/contenteditable"
     | "/_page/connectors/react"
     | "/_page/connectors/react-hook-form"
     | "/_page/connectors/tanstack-table"
@@ -413,6 +426,13 @@ declare module "@tanstack/react-router" {
       path: "/connectors/ajv";
       fullPath: "/connectors/ajv";
       preLoaderRoute: typeof PageConnectorsAjvRouteImport;
+      parentRoute: typeof PageRoute;
+    };
+    "/_page/connectors/contenteditable": {
+      id: "/_page/connectors/contenteditable";
+      path: "/connectors/contenteditable";
+      fullPath: "/connectors/contenteditable";
+      preLoaderRoute: typeof PageConnectorsContenteditableRouteImport;
       parentRoute: typeof PageRoute;
     };
     "/_page/connectors/react": {
@@ -589,6 +609,7 @@ declare module "@tanstack/react-router" {
 interface PageRouteChildren {
   PageDemosRoute: typeof PageDemosRoute;
   PageConnectorsAjvRoute: typeof PageConnectorsAjvRoute;
+  PageConnectorsContenteditableRoute: typeof PageConnectorsContenteditableRoute;
   PageConnectorsReactRoute: typeof PageConnectorsReactRoute;
   PageConnectorsReactHookFormRoute: typeof PageConnectorsReactHookFormRoute;
   PageConnectorsTanstackTableRoute: typeof PageConnectorsTanstackTableRoute;
@@ -619,6 +640,7 @@ interface PageRouteChildren {
 const PageRouteChildren: PageRouteChildren = {
   PageDemosRoute: PageDemosRoute,
   PageConnectorsAjvRoute: PageConnectorsAjvRoute,
+  PageConnectorsContenteditableRoute: PageConnectorsContenteditableRoute,
   PageConnectorsReactRoute: PageConnectorsReactRoute,
   PageConnectorsReactHookFormRoute: PageConnectorsReactHookFormRoute,
   PageConnectorsTanstackTableRoute: PageConnectorsTanstackTableRoute,